Introduce scaffolding for new JvmMetadataDetector
This introduces the basic units of the detector and the metadata.
For now, we just hide the implementation of the version detector
while we migrate the actual probe service towards the new detector.
