+++ subprojects/jvm-services/src/main/java/org/gradle/internal/jvm/inspection/DefaultJvmInstallationMetadata.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.jvm.inspection;
+
+import org.gradle.api.JavaVersion;
+
+public class DefaultJvmInstallationMetadata implements JvmInstallationMetadata {
+
+    private JavaVersion languageVersion;
+
+    public DefaultJvmInstallationMetadata(JavaVersion languageVersion) {
+        this.languageVersion = languageVersion;
+    }
+
+    @Override
+    public JavaVersion getLangageVersion() {
+        return languageVersion;
+    }
+}
